backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# File parsing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# SQL parsing
sqlparse==0.4.4

# LLM integration (optional, with fallback)
openai==1.3.0
anthropic==0.7.0

# CORS
python-multipart==0.0.6
